Move SWA continuation allocation to the NPU backend
Move the legacy SWA extend/decode implementations into a concrete NPU subclass and route ordinary NPU and DSV4 allocators through it. Introduce an overridable paged-child class factory to keep the dependency graph one-way, preserve the CUDA/HIP page-aligned SWA-tail implementation, and make the NPU tail override own its zero-prefix/-1 anchor authority.

Extraction: awk '/^    def alloc_extend\(/ {copying=1} copying && /^    def alloc_extend_swa_tail\(/ {exit} copying {print}' python/sglang/srt/mem_cache/allocator/swa.py > /private/tmp/op30-wave3-3059f067-swa-alloc_extend.before.py

Patch: apply_patch < /private/tmp/op30-wave3-3059f067-mechanical.patch

Verification: cmp /private/tmp/op30-wave3-3059f067-swa-alloc_extend.before.py /private/tmp/op30-wave3-3059f067-swa-alloc_extend.after.py; equivalent decode and normalized tail comparisons also returned zero.
