+++ python/sglang/srt/disaggregation/decode.py
@@ -1516,17 +1516,7 @@ def alloc_for_decode_prealloc(
             else torch.tensor([-1], dtype=torch.int64, device=device)
         )
         if uses_swa_tail:
-            from sglang.srt.hardware_backend.npu.allocator_npu import (
-                alloc_extend_swa_tail_npu,
-            )
-
-            kv_loc = alloc_extend_swa_tail_npu(
-                allocator,
-                prefix_lens=prefix_lens,
-                prefix_lens_cpu=prefix_lens_cpu,
-                seq_lens=seq_lens,
-                seq_lens_cpu=seq_lens_cpu,
-                last_loc=last_loc,
+            kv_loc = allocator.alloc_extend_swa_tail(
                 extend_num_tokens=alloc_fill_len,
                 swa_tail_len=swa_tail_len,
                 swa_tail_end=fill_len,
