+++ python/sglang/srt/hardware_backend/npu/allocator_npu.py
@@ -3,10 +3,16 @@ from typing import TYPE_CHECKING, Optional
 import torch
 
 from sglang.srt.mem_cache.allocator.paged import PagedTokenToKVPoolAllocator
+from sglang.srt.mem_cache.allocator.swa import (
+    SWATokenToKVPoolAllocator,
+    _is_npu,
+)
 from sglang.srt.utils import get_num_new_pages, next_power_of_2
 
+if _is_npu:
+    import torch_npu
+
 if TYPE_CHECKING:
-    from sglang.srt.mem_cache.allocator.swa import SWATokenToKVPoolAllocator
     from sglang.srt.mem_cache.memory_pool import KVCache
 
 
@@ -92,80 +98,6 @@ def get_last_loc(
         looked_up,
         torch.full_like(prefix_lens, -1),
     )
-
-
-def alloc_extend_swa_tail_npu(
-    allocator: "SWATokenToKVPoolAllocator",
-    *,
-    prefix_lens: torch.Tensor,
-    prefix_lens_cpu: torch.Tensor,
-    seq_lens: torch.Tensor,
-    seq_lens_cpu: torch.Tensor,
-    last_loc: torch.Tensor,
-    extend_num_tokens: int,
-    swa_tail_len: int,
-    swa_tail_end: int,
-) -> Optional[torch.Tensor]:
-    assert allocator.page_size > 1
-    assert len(seq_lens_cpu) == 1, "SWA tail allocation currently supports bs=1"
-    assert len(prefix_lens_cpu) == 1
-    assert 0 <= swa_tail_len <= swa_tail_end <= extend_num_tokens
-    win_start: int = swa_tail_end - swa_tail_len
-    assert win_start % allocator.page_size == 0
-
-    num_full_pages = get_num_new_pages(
-        seq_lens=seq_lens_cpu,
-        page_size=allocator.page_size,
-        prefix_lens=prefix_lens_cpu,
-    )
-    num_swa_pages = (swa_tail_len + allocator.page_size - 1) // allocator.page_size
-    if not allocator.new_pages_available(num_full_pages, num_swa_pages):
-        return None
-
-    alloc_full_indices = allocator.full_attn_allocator.alloc_extend(
-        prefix_lens,
-        prefix_lens_cpu,
-        seq_lens,
-        seq_lens_cpu,
-        last_loc,
-        extend_num_tokens,
-        num_new_pages=num_full_pages,
-    )
-    assert alloc_full_indices is not None
-
-    if swa_tail_len == 0:
-        return alloc_full_indices
-
-    device = allocator.device
-    swa_prefix_lens = torch.zeros((1,), dtype=torch.int64, device=device)
-    swa_prefix_lens_cpu = torch.zeros((1,), dtype=torch.int64)
-    swa_seq_lens = torch.tensor([swa_tail_len], dtype=torch.int64, device=device)
-    swa_seq_lens_cpu = torch.tensor([swa_tail_len], dtype=torch.int64)
-    swa_last_loc = torch.tensor([-1], dtype=torch.int64, device=device)
-
-    alloc_swa_indices = allocator.swa_attn_allocator.alloc_extend(
-        swa_prefix_lens,
-        swa_prefix_lens_cpu,
-        swa_seq_lens,
-        swa_seq_lens_cpu,
-        swa_last_loc,
-        swa_tail_len,
-        num_new_pages=num_swa_pages,
-    )
-    assert alloc_swa_indices is not None
-
-    allocator.set_full_to_swa_mapping(
-        alloc_full_indices[win_start:swa_tail_end], alloc_swa_indices
-    )
-    if win_start > 0:
-        allocator.full_to_swa_index_mapping[
-            alloc_full_indices[:win_start].to(torch.int64)
-        ] = 0
-    if swa_tail_end < extend_num_tokens:
-        allocator.full_to_swa_index_mapping[
-            alloc_full_indices[swa_tail_end:].to(torch.int64)
-        ] = 0
-    return alloc_full_indices
 
 
 class NPUPagedTokenToKVPoolAllocator(PagedTokenToKVPoolAllocator):
@@ -309,3 +241,162 @@ class NPUPagedTokenToKVPoolAllocator(PagedTokenToKVPoolAllocator):
 
         if self.debug_mode:
             assert len(torch.unique(self.free_pages)) == len(self.free_pages)
+
+
+class NPUSWATokenToKVPoolAllocator(SWATokenToKVPoolAllocator):
+    def alloc_extend(
+        self,
+        prefix_lens: torch.Tensor,
+        prefix_lens_cpu: torch.Tensor,
+        seq_lens: torch.Tensor,
+        seq_lens_cpu: torch.Tensor,
+        last_loc: torch.Tensor,  # last_loc for full layers
+        extend_num_tokens: int,
+    ):
+        assert self.page_size > 1
+
+        num_new_pages = get_num_new_pages(
+            seq_lens=seq_lens_cpu, page_size=self.page_size, prefix_lens=prefix_lens_cpu
+        )
+        if not self.new_pages_available(num_new_pages, num_new_pages):
+            return None
+
+        swa_last_loc = self.translate_loc_from_full_to_swa(last_loc)
+
+        alloc_full_indices = self.full_attn_allocator.alloc_extend(
+            prefix_lens,
+            prefix_lens_cpu,
+            seq_lens,
+            seq_lens_cpu,
+            last_loc,
+            extend_num_tokens,
+            num_new_pages=num_new_pages,
+        )
+        alloc_swa_indices = self.swa_attn_allocator.alloc_extend(
+            prefix_lens,
+            prefix_lens_cpu,
+            seq_lens,
+            seq_lens_cpu,
+            swa_last_loc,
+            extend_num_tokens,
+            num_new_pages=num_new_pages,
+        )
+        assert alloc_full_indices is not None
+        assert alloc_swa_indices is not None
+
+        self.set_full_to_swa_mapping(alloc_full_indices, alloc_swa_indices)
+
+        return alloc_full_indices
+
+    def alloc_extend_swa_tail(
+        self,
+        *,
+        extend_num_tokens: int,
+        swa_tail_len: int,
+        swa_tail_end: int,
+    ) -> Optional[torch.Tensor]:
+        allocator = self
+        device = allocator.device
+        prefix_lens_cpu = torch.tensor([0], dtype=torch.int64)
+        seq_lens_cpu = torch.tensor([extend_num_tokens], dtype=torch.int64)
+        prefix_lens = torch.tensor([0], dtype=torch.int64, device=device)
+        seq_lens = torch.tensor(
+            [extend_num_tokens], dtype=torch.int64, device=device
+        )
+        last_loc = torch.tensor([-1], dtype=torch.int64, device=device)
+        assert allocator.page_size > 1
+        assert len(seq_lens_cpu) == 1, "SWA tail allocation currently supports bs=1"
+        assert len(prefix_lens_cpu) == 1
+        assert 0 <= swa_tail_len <= swa_tail_end <= extend_num_tokens
+        win_start: int = swa_tail_end - swa_tail_len
+        assert win_start % allocator.page_size == 0
+
+        num_full_pages = get_num_new_pages(
+            seq_lens=seq_lens_cpu,
+            page_size=allocator.page_size,
+            prefix_lens=prefix_lens_cpu,
+        )
+        num_swa_pages = (swa_tail_len + allocator.page_size - 1) // allocator.page_size
+        if not allocator.new_pages_available(num_full_pages, num_swa_pages):
+            return None
+
+        alloc_full_indices = allocator.full_attn_allocator.alloc_extend(
+            prefix_lens,
+            prefix_lens_cpu,
+            seq_lens,
+            seq_lens_cpu,
+            last_loc,
+            extend_num_tokens,
+            num_new_pages=num_full_pages,
+        )
+        assert alloc_full_indices is not None
+
+        if swa_tail_len == 0:
+            return alloc_full_indices
+
+        device = allocator.device
+        swa_prefix_lens = torch.zeros((1,), dtype=torch.int64, device=device)
+        swa_prefix_lens_cpu = torch.zeros((1,), dtype=torch.int64)
+        swa_seq_lens = torch.tensor([swa_tail_len], dtype=torch.int64, device=device)
+        swa_seq_lens_cpu = torch.tensor([swa_tail_len], dtype=torch.int64)
+        swa_last_loc = torch.tensor([-1], dtype=torch.int64, device=device)
+
+        alloc_swa_indices = allocator.swa_attn_allocator.alloc_extend(
+            swa_prefix_lens,
+            swa_prefix_lens_cpu,
+            swa_seq_lens,
+            swa_seq_lens_cpu,
+            swa_last_loc,
+            swa_tail_len,
+            num_new_pages=num_swa_pages,
+        )
+        assert alloc_swa_indices is not None
+
+        allocator.set_full_to_swa_mapping(
+            alloc_full_indices[win_start:swa_tail_end], alloc_swa_indices
+        )
+        if win_start > 0:
+            allocator.full_to_swa_index_mapping[
+                alloc_full_indices[:win_start].to(torch.int64)
+            ] = 0
+        if swa_tail_end < extend_num_tokens:
+            allocator.full_to_swa_index_mapping[
+                alloc_full_indices[swa_tail_end:].to(torch.int64)
+            ] = 0
+        return alloc_full_indices
+
+    def alloc_decode(
+        self,
+        seq_lens: torch.Tensor,
+        seq_lens_cpu: torch.Tensor,
+        last_loc: torch.Tensor,  # last_loc for full layers
+    ):
+        assert self.page_size > 1
+        swa_last_loc = self.translate_loc_from_full_to_swa(last_loc)
+
+        alloc_full_indices = self.full_attn_allocator.alloc_decode(
+            seq_lens, seq_lens_cpu, last_loc
+        )
+        alloc_swa_indices = self.swa_attn_allocator.alloc_decode(
+            seq_lens, seq_lens_cpu, swa_last_loc
+        )
+
+        if alloc_full_indices is None or alloc_swa_indices is None:
+            return None
+
+        if _is_npu:
+            indices_2d = alloc_full_indices.to(torch.int64).unsqueeze(-1)
+            torch_npu.npu_scatter_nd_update_(
+                self.full_to_swa_index_mapping,
+                indices_2d,
+                alloc_swa_indices.to(torch.int64),
+            )
+        else:
+            self.full_to_swa_index_mapping[alloc_full_indices] = alloc_swa_indices
+
+        return alloc_full_indices
+
+    def _get_paged_allocator_class(
+        self,
+    ) -> type[PagedTokenToKVPoolAllocator]:
+        return NPUPagedTokenToKVPoolAllocator
